balloon: fix broken power-of-two chunk allocation
Fixes incorrect loop break condition. It was not stopping the loop when the
target number of allocated pages was reached and without that isssue it
would have stopped the loop too early, when single page allocations were still
possible.The loop is now stopped only when the current power of two is already
one and such a single page allocation has failed, or the target number of pages
to allocate has been fulfilled.

Fixes incorrect chunk sizes being allocated and logged (the exponent,
instead of the value of the power of two). This led to very small chunk
allocations with similar overhead to the old exclusively single page
allocations.