balloon: fix incorrect chunk size exponent calculation
The chunks size's exponent was being halved instead of being decreased
by one. This was because the exponent was still mistakenly being
operated on as if it were the chunk size, not the exponent of the chunks
size.
Change the naming of the variables and fix the calculation by
decrementing the exponent instead of halving it.